Refactor de los tests de gui
